requirements.txt
# Core Framework
Django>=5.0,<5.1
djangorestframework>=3.14.0,<4.0.0

# HTTP Client for GitHub API
httpx>=0.25.0,<1.0.0

# Environment Variables
python-dotenv>=1.0.0,<2.0.0

# CORS Headers
django-cors-headers>=4.3.0,<5.0.0

# API Documentation
drf-spectacular>=0.27.0,<1.0.0

# Retry and Circuit Breaker
tenacity>=8.2.0,<9.0.0
